+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+* Added `compas_occ.geometry.NurbsCurve.join`.
+* Added `compas_occ.geometry.NurbsCurve.joined`.
 
 ### Changed
+* Extend `compas_occ.geometry.OCCNurbsSurface.from_fill` with up to 4 input curves.
 
 ### Removed